Update checks for refs before doing anything

+++ src/lib/update.js
@@ -1,7 +1,13 @@
 "use strict";
 
 module.exports = function(ref, path, val) {
-    var db = path ? ref.child(path) : ref;
+    var db;
+
+    if(!ref) {
+        return;
+    }
+
+    db = path ? ref.child(path) : ref;
 
     if(val === false) {
         return db.remove();
